requirements.txt
gradio==3.35.2
tensorflow==2.12.0
numpy==1.22.4
PyYAML